learn

+++ backend/controllers/product.controller.js
@@ -38,8 +38,8 @@ export const updateProduct = async (req, res) => {
     }
 
     try {
-        const apdatedProduct = await Product.findByIdAndUpdate(id, product, { new: true })
-        res.status(200).json({ success: true, data: apdatedProduct })
+        const updatedProduct = await Product.findByIdAndUpdate(id, product, { new: true })
+        res.status(200).json({ success: true, data: updatedProduct })
     } catch (error) {
         res.status(500).json({ success: false, message: 'Server Error' })
     }
@@ -47,11 +47,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: 'Invalid Product ID' })
+    }
+
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({ success: true, message: 'Product is deleted' })
     } catch (error) {
         console.error("Error in Delete product", error.message)
-        res.status(404).json({ success: false, message: 'Product not found' })
+        res.status(500).json({ success: false, message: 'Server Error' })
     }
 }
